spool submissions report mailer
